pyproject.toml
[tool.poetry]
name = "scan"
version = "0.1.0"
description = "This project aims to develop a sophisticated multi-agent system that replicates the functions of the prefrontal cortex (PFC) using AI agents."
authors = ["John Doe <john.doe@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10, <=3.13"
crewai = "0.30.11"
duckduckgo-search = "^6.1.0"
langchain-google-genai = "^1.0.4"
python-dotenv = "^1.0.1"
tenacity = "^8.3.0"
tiktoken = ">=0.5.2,<0.6.0"
langchain-community = "0.0.25"
langchain = ">=0.1.10, <0.2.0"  # Updated version range for compatibility
langchain-openai = "^0.0.5"
chromadb = "^0.5.15"
openai = "^1.52.2"

[tool.poetry.group.dev.dependencies]
ruff = "0.4.4"
pre-commit = "^4.0.1"
mypy = {version = "^1.13.0", extras = ["faster-cache"]}

[tool.poetry.scripts]
run-scan = "scan.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
check_untyped_defs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["crewai.*"]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"
fix = true

[tool.ruff.lint]
select = [
  "E",  # pycodestyle
  "B",  # flake8-bugbear
  "W",  # Warning
  "F",  # pyflakes
  "UP",  # pyupgrade
  "I001",  # unsorted-imports
]
ignore = [
  "E501",
  "W191",
  "E111",
  "E114",
  "E117",
  "D206",
  "D300",
  "Q000",
  "Q001",
  "Q002",
  "Q003",
  "COM812",
  "COM819",
  "ISC001",
  "ISC002",
]
